requirements.txt
PyYAML==6.0.1
python-dotenv==1.0.0
numpy==1.22.4
pandas==2.0.3
python-dateutil==2.8.2
minio==7.1.15
seaborn==0.12.2
matplotlib==3.7.2
redis==4.6.0
torch==1.13.0
tensorboard==2.13.0
pymongo==4.4.1
uvicorn==0.23.1
typing_extensions==4.7.1
paramiko==3.2.0
mpi4py==3.1.4
haversine==2.8.0
antropy==0.1.6
scipy==1.11.4
